api/requirements.txt
django!=2
MySQL-python
djangorestframework